Multiple bugs fixed (ironically)
- The player can move in the Z axis

Was already fixed, but the hardcoded 0.0f in the if has been changed to
zPosition. The value of zPosition, 0.0f by default too, is now assigned
in the Start() function, so the player Z position can be moved in the
scene without any need to code.

- Camera movement

The camera used to move abruptly upon falling. The camera Y function has
been changed to one that works with the screen position rather than if
the player is jumping.  The camera will follow the player in the Y axis
when he is over the 50% of the screen and under the 33% of it (values
can be changed). This makes it move fluidly when jumping and going
underground, and gives some "secure area" of no Y movement so the player
has a chance to look in the direction he's going.

- Jumping don't work on inclinated terrain

Controls won't work properly when jumping in inclinated terrain, because
when the player moved it would enter in "Jump" state. Only being able to
jump when the button was just pressed, this would give a little window
of opportunity to actually jump.
This has been fixed by creating a new boolean that says if the jump has
actually been player-activated. It activates upon preparing jump and
deactivates anytime the key it is activated but the button is no longer
pushed. Now jumping works in "is pushed" rather than "is being pushed",
allowing the player to jump whenever is possible if he keeps pushing the
button.

- Dynamic camera zoom won't work properly

The dynamic camera final zoom would focus directly in front of it,
ignoring the player actual start position.
Now the camera uses a precalculated vector to move to its position in 2
seconds. The class functionality has been changed to a state machine for
organization reasons.

- Boxes wait time greatly reduced changing the way they work to adjust
their cooldown to the time the last box will dissapear + 1 extra second.
If cheating is a concern, the extra seconds can be added. The Box
Creator UI would need some updating.